further clearer division of synth and learn
